git-svn-id: http://localhost:9111/private/safir_open/safir_sdk_core/branches/no_ace@365 9dd7cf02-b8c9-4ff8-a4a5-e7fda3bdd973

+++ src/dose/dose_communication.ss/src/dose_com/DoseComKeepAlive.cpp
@@ -78,7 +78,7 @@ static void Handle_GetInfoRequest(CIpmSocket *pTxRxSock,
                                   DOSE_UDP_GETINFO_MSG *pRxMsg)
 {
     int     jj;
-    int     pos = 0;
+    size_t  pos = 0;
     int     result;
     dcom_ulong32 IpMulticastAddr_nw;
     dcom_ulong32 IpAddr_nw;
@@ -392,7 +392,7 @@ static void Handle_GetInfoRequest(CIpmSocket *pTxRxSock,
 
     result = pTxRxSock->SendTo2(pRxMsg->IpAddrFrom_nw,
                                 htons(pRxMsg->RespPort_nw),
-                                buff, strlen(buff), NULL, 0);
+                                buff, static_cast<unsigned long>(strlen(buff)), NULL, 0);
 }
 /*------------- end Handle_GetInfoRequest() -----------------*/
